[socket-to-api] Convert all socket events used on the landing page to API endpoints.
Actions that don't affect the contents of a workspace (like reminder emails, creating/reading/deleting workspaces) should not be handled by socket events. This is because it is clumsy for the client to have one function to emit the event, then have to listen to a different event using another function (basically a `done` callback) to handle the result.

Other things:
- Move more env vars parsing to config.js
- Slightly clean up auto-emailing logic.